requirements.txt
aio-pika==9.4.0
streamlit>=1.34.0
pandas>=2.2.0
fpdf2==2.7.6
matplotlib==3.8.0
elasticsearch[async]==8.11.0
pyyaml==6.0.1
aiohttp==3.9.0
